requirements.txt
pyTelegramBotAPI
pyscp
googlesearch-python
random2